imx: imx8mq: add SOC info SIP support for i.MX8MQ
i.MX8MQ does NOT update revision info in ANATOP_DIGPROG
register, so the revision info needs to read from ROM,
for security reason, this needs to be done in ATF, so
add this SIP support for kernel.

The A0 chip's ROM version is located at 0x800, and B0
chip is located at 0x83c.

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>