freescale: Add bakery_lock to sc_call_rpc
It is possible for multiple cores to do attempt RPC calls and without
locking this can result in corrupt messages on the SCFW side. This is
particularly visible with cpuidle.

This needs to be a bakery_lock because on the wakeup path
psci_cpu_suspend_finish calls the platform's pwr_domain_suspend_finish
before enabling the data cache.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Anson Huang <anson.huang@nxp.com>