plat: imx: add gic save/restore for imx8qm/qxp
In order to save power when AP side is suspend,
the DBLOG need to be put into low power mode.
GIC need to be power off to save power. before GIC
power off, we need to save the GIC setting, then
after resume, we need to restore the gic setting.

irqsteer need to be used to wakeup the AP side
when wakeup interrupt is pending for AP side.

Signed-off-by: Bai Ping <ping.bai@nxp.com>