console: aarch64: expose console_list for i.MX partition reboot
For i.MX SoCs with system controller inside, partition reboot
is supported, and when reboot, CPU will be reset to ATF entry
directly, the data section is NOT reset, console_list variable
needs to be initialized to 0 to make it work, otherwise, system
will be busy looping in flush_loop(). So expose console_list
for i.MX platforms to initialize this variable.

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>