plat: imx8mn: correct the noc power down setting
The NOC power down is skipped by mistake before, so fix it.
Additionally, the OCRAM memory region protection is enabled
by ROM, but does NOT disable it. it should be disabled by
default. if there are requirments for RDC memory region
protection for OCRAM space, we can reconfigure it and enable
it again.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>