imx: initialize register value before writting MU_SR register
MU_SR register bit[30] is cold boot flag passed from SCU,
w0 is random value and would clear the flag incorrectly,
and cause system partition reboot fail if Linux is in suspend.
So this patch initializes it to 0x80000000 which is exactly
the same with first time board power up before writting to MU_SR
register.

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>