MLK-24474: Add SIP call to enable FIPS mode
The configuration of the FIPS alter the SoC which is configured
and cannot be reverted so the support SHALL NOT be in
customer binary as it could lead to DoS of the SECO.

We add a SIP service to configure the FIPS mode. It is added
to the ATF because it is the only component with the required
permissions to successfully perform the call.
This service currently only allow to set the FIPS mode with
a value but can be extended.
IT can be called from other components like uboot or the OS.

The support is added only if the bl31 is compiled with
FIPS_CONFIG defined which happens when FIPS_CONFIG=on is passed
as option to Makefile.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Acked-by: Anson Huang <anson.huang@nxp.com>
Acked-by: Ye Li <ye.li@nxp.com>