LF-5427 plat:imx8 Fix missed SCU wakeup interrupts
Handle corner cases of early of wakeup interrupts.
Monitor the status of OS MU interrupt in the IRQSTR as the last step
in the suspend process. If an MU interrupt is pending, switch the
wakeup source to be irqsteer so that the core can woken up by the SCFW
after wfi is executed.

Signed-off-by: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>
(cherry picked from commit c3acc399cfd8db878d6078456092ca3f63fd070b)