MA-20469 Enable suspend/resume when GPC interrupt-controller is disabled in the kernel
Because ERR11171 and Android enable GKI, imx8mq cannot support cpuidle.
In order to support suspend/resume on imx8mq board, remove the change
of ERR11171 when IMX_ANDROID_BUILD=true. At this time, gpc monitors
the wakeup irq and can wakeup, and is consistent with the logic of
other imx8m boards.

Change-Id: I0070ff59f26ea6694c28d5b0024471d24a721bc5
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>