LF-6898 feat(imx8ulp): add pcc sw_rst workaround for lpav peripherals
For some of the peripherals in LPAV domain, the peripheral
itself may support async reset, but in SoC level, the hw reset
signal are synced to the peripheral's function clock. that means
when the peripheral is power up, the hw reset can not take effect
if the clock is disabled.

In order to reduce the changes in linux side device driver,
add sw reset assert when power up the peripheral, and the reset
will be dessert after clock is enabled. By this way,
we can meet the sync reset requirement.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>