LF-7020-1 imx8ulp: Change DRAM timing address to 0x20055000
We need to extend NOBITS and RWDATA regions for android, but as
DRAM timing space occupied about 10KB from 0x2006c000, this causes
NOBITS and RWDATA can't use it.

So move the DRAM timing base to 0x20055000 to eliminate the
limitation. DRAM timing is saved by SPL and will be saved again
by ATF when doing suspend. ATF won't touch it during booting,
so it is safe to put it in same MBC block with the image.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>