LF-7608 imx8ulp: xrdc: Skip PAC configuration for XRDC manager
When DBD owner fuse is burn to 1xx - S400 only. ATF will meet exception
when configuring the PAC 0 slot 47 which is for XRDC manager registers.
The root cause is when the word PDAC_W0 is written, the XRDC automatically
clears the valid bit in PDAC_W1. As the DBD is enabled and the valid bit is
cleared, only the DBD owners can access the XRDC manager page. However,
the A35 code continues its init sequence and tries to update PDAC_W1 that
creates an exception.

So we can just skip the configuration of this slot and use Sentinel's
setting.
Note: Sentinel FW needs update together to ensure the setting correct.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>