MA-23026 Fix bug: failed boot on LS platforms due caam driver
Recently, we enable the caam driver on i.mx platforms. But somewhat special
on i.mx because a little endian used on i.mx, and put the low bits of data
at low address space for caam's output/input address registers.
This patch consided this stituation to avoid wrong address sequence be set.

Signed-off-by: Maximus <maximus.sun@nxp.com>