MA-23473 imx93: fix memory map failed with QNX SPL
In the mmap_add API, a loop is used to add memory map regions, and this loop
terminates when it reads a 0.
Here is a memory region list that does not end with 0, which could potentially
cause the API to fail in determining the length of the memory region list.

In fact, ATF runs on OCRAM, and this memory region is set to zero by default,
so the code here can run normally.
However, when ATF is not running on OCRAM, for example, when using QNX SPL to
load it, ATF will fail to boot.

Signed-off-by: Maximus <maximus.sun@nxp.com>