LF-14964: fix(imx95): switch to forceful LM shutdown
The previous values for the 'IMX9_SCMI_LMM_SHUTDOWN_FLAG_FORCE' and
the 'IMX9_SCMI_LMM_SHUTDOWN_FLAG_GRACEFUL' flags were wrong. Because
of that, even if the 'IMX9_SCMI_LMM_SHUTDOWN_FLAG_GRACEFUL' flag was
used for the LM shutdown operation, the SM was actually doing a forceful
LM shutdown.

Since the intention was not to request the LM to shut itself down
(via the graceful shutdown flag), change to using the FORCE shutdown
flag.

Change-Id: I206edd166c7a8b760143da43f4576a6ebbf4f554
Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>