LF-16690: imx93: add trdc config for NSE access to 384K OCRAM
The initial 384K of OCRAM is available for non secure world to use.
Current setting in TRDC configuration for NSE access is only for 96K,
extend it to cover the rest od 384K space.

This would allow user to read/write this memory from u-boot prompt.
Usecase targetted is TFM enablement on Cortex M core where non-secure
binary for Cortex M is loaded on OCRAM.

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>